🎨: デザインシステム連携のため余白ダイアログパターンの記述構成を見直す
